+++ portfolio/lib/app/datamodels/coin.dart
@@ -1,0 +1,7 @@
+class Coin {
+  int id;
+  String symbol;
+  String name;
+
+  Coin({this.id, this.symbol, this.name});
+}
