+++ portfolio/lib/app/models/coin_model.dart
@@ -1,0 +1,9 @@
+import 'package:portfolio/app/datamodels/coin.dart';
+import 'package:portfolio/core/base_viewmodel.dart';
+
+class CoinModel extends BaseViewModel {
+  List<Coin> _coins = [];
+  List<Coin> get coins => _coins;
+
+  Future loadCoins() {}
+}
