+++ portfolio/lib/app/models/portfolio_model.dart
@@ -1,8 +1,8 @@
-import 'package:portfolio/core/datamodels/porfolio.dart';
+import 'package:portfolio/app/datamodels/porfolio.dart';
+import 'package:portfolio/app/ui/services/database_service.dart';
+import 'package:portfolio/core/base_viewmodel.dart';
 import 'package:portfolio/core/enums/viewstate.dart';
-import 'package:portfolio/core/services/database_service.dart';
 import 'package:portfolio/locator.dart';
-import 'package:portfolio/ui/views/base/base_viewmodel.dart';
 
 class PortfolioModel extends BaseViewModel {
   List<Portfolio> _portfolios = [];
