--- portfolio/lib/core/datamodels/coin.dart
@@ -1,6 +1,0 @@
-class Coin {
-  int id;
-  String symbol;
-
-  Coin({this.id, this.symbol});
-}
