+++ portfolio/lib/locator.dart
@@ -1,10 +1,10 @@
 import 'package:get_it/get_it.dart';
-import 'package:portfolio/core/changenotifiers/portfolio_model.dart';
-import 'package:portfolio/ui/views/dashboard/dashboard_viewmodel.dart';
-import 'package:portfolio/ui/views/startup/startup_viewmodel.dart';
+import 'package:portfolio/app/models/coin_model.dart';
+import 'package:portfolio/app/models/portfolio_model.dart';
+import 'package:portfolio/app/ui/services/database_service.dart';
+import 'package:portfolio/app/ui/views/dashboard/dashboard_viewmodel.dart';
+import 'package:portfolio/app/ui/views/startup/startup_viewmodel.dart';
 import 'package:stacked_services/stacked_services.dart';
-
-import 'core/services/database_service.dart';
 
 GetIt locator = GetIt.instance;
 
@@ -17,4 +17,5 @@ void setupLocator() {
   locator.registerLazySingleton(() => DashboardViewModel());
   locator.registerLazySingleton(() => StartupViewModel());
   locator.registerLazySingleton(() => PortfolioModel());
+  locator.registerLazySingleton(() => CoinModel());
 }
